backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
gunicorn>=22.0
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
reportlab>=4.1
openpyxl>=3.1
python-docx>=1.1
